// project/pc-vue3-ts/src/stores/base/system/type.ts
export interface IUser {
  id: number
  name: string
  realname: string
  cellphone: number
  enable: number
  departmentId: number
  roleId: number
  createAt: string
  updateAt: string
}

export interface ISystemState {
  usersTotalCount: number
  usersList: IUser[]
  pageList: any[]
  pageTotalCount: number
  dictTree: any[]
}